+++ src/Types.fs
@@ -1,6 +1,34 @@
 namespace Kafka
 
 //
+// Common
+//
+
+type BrokerList = BrokerList of string
+type StreamName = StreamName of string
+
+module StreamName =
+    let value (StreamName streamName) = streamName
+
+type GroupId =
+    | Random
+    | Id of string
+
+module GroupId =
+    let map f = function
+        | Id groupId -> groupId |> f |> Id
+        | Random -> Random
+
+    let value = function
+        | Id groupId -> groupId
+        | Random -> sprintf "random-%d" System.DateTime.Now.Ticks    // random (unique) group id means, it will always starts from the beginning
+
+type Configuration = {
+    BrokerList: BrokerList
+    Topic: StreamName
+}
+
+//
 // Kafka readers
 //
 
@@ -16,8 +44,3 @@ type ParsedMessageReader<'Event> = {
 type MessageReader<'Event> =
     | DecodedMessageReader of DecodedMessageReader
     | ParsedMessageReader of ParsedMessageReader<'Event>
-
-type Configuration = {
-    BrokerList: string
-    Topic: string
-}