+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Add more common types
+    - `BrokerList`
+    - `StreamName`
+    - `GroupId`
 
 ## 4.4.0 - 2019-04-04
 - Add `Admin` module with simple topic meta information